Update user-agent.rkt
implementation of domain-match? ignores the first part of 5.3.1 of RFC 6265, which reads " The domain string and the string are identical.  (Note that both the domain string and the string will have been canonicalized to lower case at this point.)".  Rather, the function only checked the the second condition was satisfied.  This caused an error where ip-address hosts would not match to themselves.