make-cookie should coerce name and value to strings
Currently, the contract for make-cookie accepts its name and value
args as either strings or byte-strings, but the contracts for
`cookie-name` and `cookie-value` require strings.

This means that `(define c (make-cookie #"name" #"val")` succeeds,
but `(cookie-name c)` raises an error.

This is fixed by having make-cookie convert its arguments as needed.